requirements.txt
gspread
google-auth
fastapi
uvicorn
